🎨 move mock list into HabitList page
