+++ src/app/App.module.css
@@ -40,7 +40,3 @@
 button {
   font-size: calc(10px + 2vmin);
 }
-
-.habitColor {
-  height: 2em;
-}
