+++ src/app/pages/HabitList/HabitList.tsx
@@ -4,7 +4,55 @@ import PageLink from '../../components/PageLink/PageLink';
 import SwipeDate from '../../components/SwipeDate/SwipeDate';
 import styles from './HabitList.module.css';
 
+type Habit = {
+  id: string;
+  name: string;
+  color: string;
+  dateCreated: string;
+  datesCompleted: string[];
+  rating: string;
+};
+
+function toListElement(habit: Habit): JSX.Element {
+  return (
+    <li key={habit.id}>
+      <input type="checkbox" /> {habit.name}
+      <div
+        className={styles.habitColor}
+        style={{ backgroundColor: habit.color }}
+      ></div>
+    </li>
+  );
+}
+
 function HabitList(): JSX.Element {
+  const habits: Habit[] = [
+    {
+      id: '1',
+      name: 'take a break',
+      color: 'hotpink',
+      dateCreated: '12.3.21',
+      datesCompleted: [],
+      rating: 'nice',
+    },
+    {
+      id: '2',
+      name: 'take a break',
+      color: 'hotpink',
+      dateCreated: '12.3.21',
+      datesCompleted: [],
+      rating: 'nice',
+    },
+    {
+      id: '3',
+      name: 'take a break',
+      color: 'hotpink',
+      dateCreated: '12.3.21',
+      datesCompleted: [],
+      rating: 'nice',
+    },
+  ];
+
   return (
     <div className={styles.container}>
       <header>
@@ -12,6 +60,7 @@ function HabitList(): JSX.Element {
       </header>
       <main className={styles.main}>
         <SwipeDate />
+        <ul>{habits.map(toListElement)}</ul>
 
         <PageLink variant="plus" link="/create-habit">
           <PlusIcon />
